@@ -4,7 +4,7 @@
   "description": "Personal finance app",
   "main": "index.js",
   "engines": {
-    "node": "8.9.x"
+    "node": "10"
   },
   "sasslintConfig": ".sass-lint.yml",
   "scripts": {
@@ -13,6 +13,7 @@
     "dev": "NODE_ENV=development node index.js",
     "dev:api": "NODE_ENV=development SKIP_APP=true nodemon --inspect -w ./api/src index.js",
     "test": "NODE_ICU_DATA=node_modules/full-icu NODE_ENV=test mocha --require babel-core/register --require ignore-styles -R spec \"./web/test/**/*.spec.js\" \"./api/test/**/*.spec.js\"",
+    "test:dev": "NODE_ICU_DATA=node_modules/full-icu NODE_ENV=test mocha --require babel-core/register --require ignore-styles -R spec -w --bail \"./web/test/**/*.spec.js\" \"./api/test/**/*.spec.js\"",
     "lint": "eslint webpack.config.js webpack api/src api/test web/src web/test",
     "sass-lint": "sass-lint -qv \"./web/src/sass/**/*.scss\"",
     "scrape_funds": "node ./api/scripts/scrapeFunds/index.js --prices",
@@ -78,8 +79,9 @@
     }
   },
   "dependencies": {
+    "@hot-loader/react-dom": "^16.8.4",
     "axios": "^0.16.2",
-    "bcrypt": "^1.0.3",
+    "bcrypt": "^2.0.1",
     "body-parser": "^1.17.2",
     "classnames": "^2.2.5",
     "color-hash": "^1.0.3",
@@ -112,8 +114,8 @@
     "postcss-import": "^11.0.0",
     "postcss-loader": "^2.0.9",
     "prompt": "^1.0.0",
-    "react": "^16.4.0",
-    "react-dom": "^16.4.0",
+    "react": "^16.8.4",
+    "react-dom": "^16.8.4",
     "react-immutable-proptypes": "^2.1.0",
     "react-media": "^1.6.1",
     "react-pure-render": "^1.0.2",
@@ -142,8 +144,8 @@
     "css-loader": "^0.28.11",
     "deep-equal": "^1.0.1",
     "dotenv": "^5.0.1",
-    "enzyme": "^3.3.0",
-    "enzyme-adapter-react-16": "^1.1.1",
+    "enzyme": "^3.9.0",
+    "enzyme-adapter-react-16": "^1.10.0",
     "eslint": "^4.5.0",
     "eslint-plugin-babel": "^4.1.2",
     "eslint-plugin-react": "^7.3.0",
@@ -160,7 +162,8 @@
     "mock-knex": "^0.4.0",
     "node-sass": "^4.9.0",
     "nodemon": "^1.11.0",
-    "react-hot-loader": "^4.2.0",
+    "react-hot-loader": "^4.8.0",
+    "react-testing-library": "^6.0.0",
     "redux-logger": "^3.0.6",
     "redux-saga-test-plan": "^3.3.1",
     "redux-test-utils": "^0.2.2",
@@ -172,7 +175,7 @@
     "sinon-chai": "^2.14.0",
     "style-loader": "^0.19.1",
     "webpack": "^4.8.3",
-    "webpack-cli": "^2.1.4",
+    "webpack-cli": "^2.1.5",
     "webpack-dev-middleware": "^3.1.3",
     "webpack-hot-middleware": "^2.22.2"
   }
