@@ -1,13 +1,27 @@
 import React from 'react';
 import PropTypes from 'prop-types';
 
-export default function FormFieldCost({ value, onChange }) {
-    const procValue = value / 100;
-    const procOnChange = evt => onChange(Math.round(100 * Number(evt.target.value)));
+import { useField } from './use-field';
 
-    return <div className="form-field form-field-cost">
-        <input type="number" step="0.01" defaultValue={procValue} onChange={procOnChange} />
-    </div>;
+export default function FormFieldCost({ value, onChange }) {
+    const [currentValue, onType, onBlur] = useField(
+        value,
+        onChange,
+        cost => cost || 0,
+        cost => Math.round(100 * Number(cost))
+    );
+
+    return (
+        <div className="form-field form-field-cost">
+            <input
+                type="number"
+                step="0.01"
+                defaultValue={currentValue / 100}
+                onChange={onType}
+                onBlur={onBlur}
+            />
+        </div>
+    );
 }
 
 FormFieldCost.propTypes = {
