@@ -3,13 +3,26 @@ import PropTypes from 'prop-types';
 import { DateTime } from 'luxon';
 import { dateInput } from '../../helpers/date';
 
-export default function FormFieldDate({ value, onChange }) {
-    const procValue = value.toISODate();
-    const procOnChange = evt => onChange(dateInput(evt.target.value, false));
+import { useField } from './use-field';
 
-    return <div className="form-field form-field-date">
-        <input type="date" defaultValue={procValue} onChange={procOnChange} />
-    </div>;
+export default function FormFieldDate({ value, onChange }) {
+    const [currentValue, onType, onBlur] = useField(
+        value,
+        onChange,
+        date => dateInput(date, false),
+        date => dateInput(date, false)
+    );
+
+    return (
+        <div className="form-field form-field-date">
+            <input
+                type="date"
+                defaultValue={currentValue.toISODate()}
+                onChange={onType}
+                onBlur={onBlur}
+            />
+        </div>
+    );
 }
 
 FormFieldDate.propTypes = {
