@@ -1,12 +1,26 @@
 import React from 'react';
 import PropTypes from 'prop-types';
 
-export default function FormFieldText({ value, onChange }) {
-    const procOnChange = evt => onChange(evt.target.value);
+import { useField } from './use-field';
 
-    return <div className="form-field form-field-text">
-        <input type="text" defaultValue={value} onChange={procOnChange} />
-    </div>;
+export default function FormFieldText({ value, onChange }) {
+    const [currentValue, onType, onBlur] = useField(
+        value,
+        onChange,
+        text => text,
+        text => text
+    );
+
+    return (
+        <div className="form-field form-field-text">
+            <input
+                type="text"
+                defaultValue={currentValue}
+                onChange={onType}
+                onBlur={onBlur}
+            />
+        </div>
+    );
 }
 
 FormFieldText.propTypes = {
