@@ -1,54 +1,49 @@
 import { Map as map } from 'immutable';
 import { PAGES } from '../../constants/data';
-import React from 'react';
-import ImmutableComponent from '../../ImmutableComponent';
+import React, { useState, useCallback, useMemo } from 'react';
 import PropTypes from 'prop-types';
 import classNames from 'classnames';
 import GraphFundItem from '../GraphFundItem';
 import FundGainInfo from '../FundGainInfo';
 
-export default class ListRowFundsDesktop extends ImmutableComponent {
-    static propTypes = {
-        row: PropTypes.instanceOf(map).isRequired
-    };
-    constructor(props) {
-        super(props);
+const itemKey = PAGES.funds.cols.indexOf('item');
 
-        this.state = {
-            popout: false
-        };
-    }
-    onToggleGraph() {
-        this.setState({ popout: !this.state.popout });
-    }
-    render() {
-        const { row } = this.props;
-        const { popout } = this.state;
+export default function ListRowFundsDesktop({ row }) {
+    const [popout, setPopout] = useState(false);
 
-        const itemKey = PAGES.funds.cols.indexOf('item');
-        const name = row.getIn(['cols', itemKey])
+    const onToggleGraph = useCallback(() => {
+        setPopout(!popout);
+    });
+
+    const name = useMemo(
+        () => row.getIn(['cols', itemKey])
             .toLowerCase()
-            .replace(/\W+/g, '-');
+            .replace(/\W+/g, '-'),
+        [row]
+    );
 
-        const sold = row.get('sold');
+    const sold = row.get('sold');
 
-        const className = classNames('fund-extra-info', { popout });
+    const className = classNames('fund-extra-info', { popout });
 
-        return (
-            <span className={className}>
-                <span className="fund-graph">
-                    <div className="fund-graph-cont">
-                        <GraphFundItem name={name}
-                            sold={sold}
-                            values={row.get('prices')}
-                            popout={popout}
-                            onToggle={() => this.onToggleGraph()}
-                        />
-                    </div>
-                </span>
-                <FundGainInfo gain={row.get('gain')} sold={sold} />
+    return (
+        <span className={className}>
+            <span className="fund-graph">
+                <div className="fund-graph-cont">
+                    <GraphFundItem name={name}
+                        sold={sold}
+                        values={row.get('prices')}
+                        popout={popout}
+                        onToggle={onToggleGraph}
+                    />
+                </div>
             </span>
-        );
-    }
+            <FundGainInfo gain={row.get('gain')} sold={sold} />
+        </span>
+    );
 }
+
+ListRowFundsDesktop.propTypes = {
+    row: PropTypes.instanceOf(map).isRequired
+};
 
