@@ -1,5 +1,7 @@
+import '../../browser';
 import { expect } from 'chai';
-import { shallow } from 'enzyme';
+import 'react-testing-library/cleanup-after-each';
+import { render, fireEvent } from 'react-testing-library';
 import React from 'react';
 import FormFieldNumber from '../../../src/components/FormField/number';
 
@@ -15,28 +17,39 @@ describe('<FormFieldNumber />', () => {
     };
 
     it('should render its basic structure', () => {
-        const wrapper = shallow(<FormFieldNumber {...props} />);
+        const { container } = render(<FormFieldNumber {...props} />);
 
-        expect(wrapper.is('div.form-field.form-field-number')).to.equal(true);
-        expect(wrapper.children()).to.have.length(1);
+        const [div] = container.childNodes;
+
+        expect(div.tagName).to.equal('DIV');
+        expect(div.className).to.equal('form-field form-field-number');
+        expect(div.childNodes).to.have.length(1);
     });
 
     it('should render an input', () => {
-        const wrapper = shallow(<FormFieldNumber {...props} />);
+        const { container } = render(<FormFieldNumber {...props} />);
 
-        expect(wrapper.childAt(0).is('input')).to.equal(true);
-        expect(wrapper.childAt(0).props()).to.deep.include({
-            type: 'number',
-            defaultValue: 103
-        });
+        const [div] = container.childNodes;
+        const [input] = div.childNodes;
+
+        expect(input.tagName).to.equal('INPUT');
+        expect(input.type).to.equal('number');
+        expect(input.value).to.equal('103');
     });
 
     it('should fire onChange', () => {
-        const wrapper = shallow(<FormFieldNumber {...props} />);
+        const { container } = render(<FormFieldNumber {...props} />);
+
+        const [div] = container.childNodes;
+        const [input] = div.childNodes;
 
         expect(changed).to.equal(null);
 
-        wrapper.childAt(0).simulate('change', { target: { value: '10.93' } });
+        fireEvent.change(input, { target: { value: '10.93' } });
+
+        expect(changed).to.equal(null);
+
+        fireEvent.blur(input);
 
         expect(changed).to.equal(10.93);
     });
