@@ -1,33 +1,19 @@
 /* eslint-disable newline-per-chained-call */
-import { expect } from 'chai';
+import '../../browser';
+import chai, { expect } from 'chai';
+import sinon from 'sinon';
+import sinonChai from 'sinon-chai';
+chai.use(sinonChai);
 import itEach from 'it-each';
 itEach();
-import { shallow } from 'enzyme';
+import 'react-testing-library/cleanup-after-each';
+import { render, fireEvent } from 'react-testing-library';
+import { DateTime } from 'luxon';
 import React from 'react';
 import FormFieldTransactions from '../../../src/components/FormField/transactions';
 import { TransactionsList } from '../../../src/helpers/data';
-import { dateInput } from '../../../src/helpers/date';
-import FormFieldDate from '../../../src/components/FormField/date';
-import FormFieldNumber from '../../../src/components/FormField/number';
-import FormFieldCost from '../../../src/components/FormField/cost';
 
 describe('<FormFieldTransactions />', () => {
-    let date = null;
-    let units = null;
-    let cost = null;
-
-    const onDateChange = (value, _ymd, key) => {
-        date = { value, ymd: _ymd, key };
-    };
-
-    const onUnitsChange = (value, _units, key) => {
-        units = { value, units: _units, key };
-    };
-
-    const onCostChange = (value, _cost, key) => {
-        cost = { value, cost: _cost, key };
-    };
-
     const transactions = [
         { date: '2017-11-10', units: 10.5, cost: 50 },
         { date: '2018-09-05', units: -3, cost: -40 }
@@ -35,94 +21,172 @@ describe('<FormFieldTransactions />', () => {
 
     const value = new TransactionsList(transactions);
 
-    const props = {
-        value,
-        onDateChange,
-        onUnitsChange,
-        onCostChange
-    };
-
     it('should render its basic structure', () => {
-        const wrapper = shallow(<FormFieldTransactions {...props} />);
+        const onChange = sinon.spy();
+        const { container } = render(<FormFieldTransactions
+            value={value}
+            onChange={onChange}
+        />);
 
-        expect(wrapper.is('ul.transactions-list')).to.equal(true);
-        expect(wrapper.children()).to.have.length(2);
+        const [ul] = container.childNodes;
+
+        expect(ul.tagName).to.equal('UL');
+        expect(ul.className).to.equal('transactions-list');
+
+        expect(ul.childNodes).to.have.length(2);
     });
 
-    let key = 0;
+    let index = null;
 
     beforeEach(() => {
-        date = null;
-        units = null;
-        cost = null;
+        index = 0;
     });
 
-    // eslint-disable-next-line max-statements
-    it.each(transactions, 'should render a list of transactions', ({
-        date: iDate, units: iUnits, cost: iCost
-    }) => {
+    it.each(transactions, 'should render a list of transactions', () => {
+        const onChange = sinon.spy();
 
-        const wrapper = shallow(<FormFieldTransactions {...props} />);
+        const { container } = render(<FormFieldTransactions
+            value={value}
+            onChange={onChange}
+        />);
 
-        expect(wrapper.childAt(key).is('li')).to.equal(true);
-        expect(wrapper.childAt(key).children()).to.have.length(1);
-        expect(wrapper.childAt(key).childAt(0).is('span.transaction')).to.equal(true);
-        expect(wrapper.childAt(key).childAt(0).children()).to.have.length(3);
+        const [ul] = container.childNodes;
+        const li = ul.childNodes[index];
 
-        // test the date input
-        expect(wrapper.childAt(key).childAt(0).childAt(0).is('span.row')).to.equal(true);
-        expect(wrapper.childAt(key).childAt(0).childAt(0).children()).to.have.length(2);
-        expect(wrapper.childAt(key).childAt(0).childAt(0).childAt(0).is('span.col')).to.equal(true);
-        expect(wrapper.childAt(key).childAt(0).childAt(0).childAt(0).text()).to.equal('Date:');
-        expect(wrapper.childAt(key).childAt(0).childAt(0).childAt(1).is('span.col')).to.equal(true);
-        expect(wrapper.childAt(key).childAt(0).childAt(0).childAt(1).children()).to.have.length(1);
-        expect(wrapper.childAt(key).childAt(0).childAt(0).childAt(1).childAt(0).is(FormFieldDate))
-            .to.equal(true);
+        expect(li.tagName).to.equal('LI');
+        expect(li.childNodes).to.have.length(1);
 
-        expect(wrapper.childAt(key).childAt(0).childAt(0).childAt(1).childAt(0).props().value
-            .toISODate())
-            .to.equal(iDate);
+        const [transaction] = li.childNodes;
 
-        expect(date).to.equal(null);
-        wrapper.childAt(key).childAt(0).childAt(0).childAt(1).childAt(0).props().onChange(dateInput('5/4/16'));
-        expect(date).to.deep.equal({ value, ymd: dateInput('5/4/16'), key });
+        expect(transaction.tagName).to.equal('SPAN');
+        expect(transaction.childNodes).to.have.length(3);
 
-        // test the units input
-        expect(wrapper.childAt(key).childAt(0).childAt(1).is('span.row')).to.equal(true);
-        expect(wrapper.childAt(key).childAt(0).childAt(1).children()).to.have.length(2);
-        expect(wrapper.childAt(key).childAt(0).childAt(1).childAt(0).is('span.col')).to.equal(true);
-        expect(wrapper.childAt(key).childAt(0).childAt(1).childAt(0).text()).to.equal('Units:');
-        expect(wrapper.childAt(key).childAt(0).childAt(1).childAt(1).is('span.col')).to.equal(true);
-        expect(wrapper.childAt(key).childAt(0).childAt(1).childAt(1).children()).to.have.length(1);
-        expect(wrapper.childAt(key).childAt(0).childAt(1).childAt(1).childAt(0).is(FormFieldNumber))
-            .to.equal(true);
-        expect(wrapper.childAt(key).childAt(0).childAt(1).childAt(1).childAt(0).props())
-            .to.have.property('value', iUnits);
+        transaction.childNodes.forEach(row => {
+            expect(row.tagName).to.equal('SPAN');
+            expect(row.className).to.equal('row');
+        });
 
-        expect(units).to.equal(null);
-        wrapper.childAt(key).childAt(0).childAt(1).childAt(1).childAt(0).props().onChange(1000);
-        expect(units).to.deep.equal({ value, units: 1000, key });
+        index++;
+    });
 
-        // test the units input
-        expect(wrapper.childAt(key).childAt(0).childAt(2).is('span.row')).to.equal(true);
-        expect(wrapper.childAt(key).childAt(0).childAt(2).children()).to.have.length(2);
-        expect(wrapper.childAt(key).childAt(0).childAt(2).childAt(0).is('span.col')).to.equal(true);
-        expect(wrapper.childAt(key).childAt(0).childAt(2).childAt(0).text()).to.equal('Cost:');
-        expect(wrapper.childAt(key).childAt(0).childAt(2).childAt(1).is('span.col')).to.equal(true);
-        expect(wrapper.childAt(key).childAt(0).childAt(2).childAt(1).children()).to.have.length(1);
-        expect(wrapper.childAt(key).childAt(0).childAt(2).childAt(1).childAt(0).is(FormFieldCost))
-            .to.equal(true);
-        expect(wrapper.childAt(key).childAt(0).childAt(2).childAt(1).childAt(0).props())
-            .to.have.property('value', iCost);
+    it.each(transactions, 'handle date input', () => {
+        const onChange = sinon.spy();
 
-        expect(cost).to.equal(null);
-        wrapper.childAt(key).childAt(0).childAt(2).childAt(1).childAt(0).props().onChange(2000);
-        expect(cost).to.deep.equal({ value, cost: 2000, key });
+        const { container } = render(<FormFieldTransactions
+            value={value}
+            onChange={onChange}
+        />);
 
-        key++;
+        const [ul] = container.childNodes;
+        const li = ul.childNodes[index];
+        const [transaction] = li.childNodes;
 
-        date = null;
-        units = null;
-        cost = null;
+        const [dateRow] = transaction.childNodes;
+
+        expect(dateRow.childNodes).to.have.length(2);
+        const [dateLabelCol, dateInputCol] = dateRow.childNodes;
+
+        expect(dateLabelCol.tagName).to.equal('SPAN');
+        expect(dateLabelCol.className).to.equal('col');
+        expect(dateLabelCol.innerHTML).to.equal('Date:');
+
+        expect(dateInputCol.tagName).to.equal('SPAN');
+        expect(dateInputCol.className).to.equal('col');
+        expect(dateInputCol.childNodes).to.have.length(1);
+
+        const { childNodes: [inputDate] } = dateInputCol.childNodes[0];
+
+        fireEvent.change(inputDate, { target: { value: '2017-04-03' } });
+        fireEvent.blur(inputDate);
+
+        expect(onChange).to.have.been.calledWith(
+            value.list,
+            index,
+            DateTime.fromISO('2017-04-03'),
+            'date'
+        );
+
+        index++;
+    });
+
+    it.each(transactions, 'handle units input', () => {
+        const onChange = sinon.spy();
+
+        const { container } = render(<FormFieldTransactions
+            value={value}
+            onChange={onChange}
+        />);
+
+        const [ul] = container.childNodes;
+        const li = ul.childNodes[index];
+        const [transaction] = li.childNodes;
+
+        const [, unitsRow] = transaction.childNodes;
+
+        expect(unitsRow.childNodes).to.have.length(2);
+        const [unitsLabelCol, unitsInputCol] = unitsRow.childNodes;
+
+        expect(unitsLabelCol.tagName).to.equal('SPAN');
+        expect(unitsLabelCol.className).to.equal('col');
+        expect(unitsLabelCol.innerHTML).to.equal('Units:');
+
+        expect(unitsInputCol.tagName).to.equal('SPAN');
+        expect(unitsInputCol.className).to.equal('col');
+        expect(unitsInputCol.childNodes).to.have.length(1);
+
+        const { childNodes: [inputUnits] } = unitsInputCol.childNodes[0];
+
+        fireEvent.change(inputUnits, { target: { value: '34.2219' } });
+        fireEvent.blur(inputUnits);
+
+        expect(onChange).to.have.been.calledWith(
+            value.list,
+            index,
+            34.2219,
+            'units'
+        );
+
+        index++;
+    });
+
+    it.each(transactions, 'handle cost input', () => {
+        const onChange = sinon.spy();
+
+        const { container } = render(<FormFieldTransactions
+            value={value}
+            onChange={onChange}
+        />);
+
+        const [ul] = container.childNodes;
+        const li = ul.childNodes[index];
+        const [transaction] = li.childNodes;
+
+        const [, , costRow] = transaction.childNodes;
+
+        expect(costRow.childNodes).to.have.length(2);
+        const [costLabelCol, costInputCol] = costRow.childNodes;
+
+        expect(costLabelCol.tagName).to.equal('SPAN');
+        expect(costLabelCol.className).to.equal('col');
+        expect(costLabelCol.innerHTML).to.equal('Cost:');
+
+        expect(costInputCol.tagName).to.equal('SPAN');
+        expect(costInputCol.className).to.equal('col');
+        expect(costInputCol.childNodes).to.have.length(1);
+
+        const { childNodes: [inputcost] } = costInputCol.childNodes[0];
+
+        fireEvent.change(inputcost, { target: { value: '126.7692' } });
+        fireEvent.blur(inputcost);
+
+        expect(onChange).to.have.been.calledWith(
+            value.list,
+            index,
+            12677,
+            'cost'
+        );
+
+        index++;
     });
 });
+
